pyproject.toml
[project]
name = "iqvara-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-ckeditor-5>=0.2.18",
    "django-debug-toolbar>=6.0.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.11",
]
